requirements.txt
kaggle==1.6.13
pandas==2.2.2
numpy==1.26.4
textblob==0.17.1
nltk==3.8.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
kaggle